storage: introduce garbage collector
Garbage collector is a special background fiber, which periodically
wakes up. The fiber incrementaly deletes garbage bucket's data
from each space in little portions (1000 tuples per time). Garbage
collector works on master only.

Garbage collection is applied to only sharded spaces. Space is
considered to be sharded, if
* it has unsigned 'bucket_id' field in format;
* it has index, which first part is the 'bucket_id' field.

Bucket is garbage if it is not in _bucket space or if its state
is 'sent', or if it is in 'garbage' state. Sent buckets are
deleted from _bucket after a specified duration.

Closes #38
Closes #15