storage: introduce rebalancer
Rebalancer is a background fiber, single per cluster and working
on an active replicaset master with the smallest UUID. It sends
requests on buckets sending to all masters.

Bucket routes are calculated in the best way to reach maximal
balance.

User can specify maximal disbslance threshold, rebalancer's idle
and work timeout. If a balance is reached, the rebalancer becames
idle, and wakeups rarely, one time per idle timeout. If a rebalance
is in progress, then the rebalancer wakeups one time per work timeout.

Each replicaset has weight - this value determines bucket count relative
to other replicasets.

A count of buckets, receiving by a replicaset at once is limited, because
else a storm of incomming buckets can appear, and it downs a server.

Closes #3
Closes #4
Closes #40