Check UUID correctness on storage and routers after connection
If a router or a storage connects to another storage and sees
UUID different from a specified in config, then there are configuration
problems. Router must not use the invalid replicaset.

If an UUID of connection is changed after reconnect (for example,
on the same port another instance has started), then log error
and close connection.

Closes #7