rebalancer: do not use lua state to check rebalancing progress
Lua flags, set in a function and unset out of it, are not reliable
way to store a state because of lua errors like OOM. If a function
fails with one of such errors, the flag can not be unsed and the
state became inconsistent.

Do not use flags to check, if a rebalancing is in progress. Instead of
it, store a fiber which applies routes, and check its status.

Closes #3