router: introduce failover
Failover connection is a connection to the nearest available server.
Failover connection is hold for each replicaset. These connections
in common case are more optimal than master connections, and can be
used for read-only queries.

The patch introduces a special fiber on router to maintain failover
connections using weight matrix from config.

Failover fiber periodically scans replicasets for disconnected or not
optimal failover connections. If a failover connection is down too long,
it tries to connect to the replica with lower priority.

If a failover connection is not optimal (has not the highest priority in
weight matrix), the fiber periodically tries to reconnect to an optimal
replica.

Closes #12