Recreate replicaset objects on reconfig
Now on reconfig old connection objects are moved into new replica objects.
But netbox connection stores two sharding attributes: replica and replicaset.
When a connection is moved, it is not updates its attributes that is wrong.

To avoid such errors lets better recreate not only replicasets, but
old connections too.

Closes #66