router: ping replicas to detect disconnection on power off
If a host is down by power off, then it does not manage to close
connection, and a router thinks, that the connection is ok even if
catches many many timeouts.

Lets detect these disconnections by periodic ping of replicas. And if a
ping does not response in 5 seconds, then consider this connection as
closed.