Introduce generic configuration validator
The more constants become options, the more duplicated code
appears in cfg.lua to validate options and in
storage(router)/init.lua to fill default values.

Allow to specify these options in congiguration templates like it
is done in Tarantool for box.cfg, create_space, create_index ...
And fill optional and absent values with default ones specified
in a template.

Closes #78