router: connect to all replicas including slaves
There are two reasons. First - if a master is changed on a
replicaset, then a router must immediately switch to it all
'write' requests. Now to do it a router creates a new connection
to a new master.

Second reason: #75. To fast detect master role change on a
replica a router must be connected to all replicas of a
replicaset. Else to establish connections to slaves the router
will spend to many time.

Closes #76