router: fix router.info bucket statistics
Router discovery fiber periodically fetches bucket identifiers of
each replicaset. If it is done during rebalancing, than it is
possible, that a bucket replicaset is changed, and router must
not only update routing map, but also update local bucket_count
of both old and new replicasets. The old one must be decremented,
and the new one must be incremented. New is just reset to the
total count now, but the old one was not updated before the
patch. It leads to a strange thing, when a bucket is accounted
twice until next discovery round.

The fix is trivial - just do decrement local bucket count of the
old replicaset.

Closes #97