storage: expose public API for getting sharded spaces
It is possible that a user makes wrong configuration of a sharded
space, for example, by forgetting bucket_id index, or making
bucket identifier field by non-number (any, string, scalar ...).
It leads to incorrect rebalancing, when _bucket is rebalanced ok,
but space is not. To simplify debugging a user application it can
be useful to open the find_sharded_spaces as a public API.

Closes #96