router: fix bucket_set to do not decrement bc always
Consider this situation: two fibers sends a request on the same
bucket and can not find it in route_map. They both start
discovery. When bucket is discovered, both fibers decreases old
bucket_count making it negative, and then increment if not equal.
It leads to negative bucket counts.

On route map update it is necessary to check old and new
replicaset on equality.

Closes #108