recovery: do not recovery a bucket beeing sent now
During rebalancing the recoverer can make a bucket be ACTIVE on
two replicasets. It is possible, if bucket_send during marking a
bucket to be SENDING does too long WAL write. During the write
the recovery fiber wakes up, checks that the bucket does not
exist on the destination replicaset and makes it ACTIVE on the
source. Then bucket_send finalizes WAL write and sends data. If
the data is applied too long on the destination, bucket_send
catches TIMEOUT error and exits. So now the bucket is ACTIVE both
on the source and the destination.

The fix is just do not touch beeing sent bucket in the recovery
fiber.

Closes #113