gc: make bucket_generation be module scope parameter
Bucket generation is a number that is incremented on each _bucket
change and is used by bucket GC to detect that _bucket is changed
even since a previous GC. But it will be usefull also to insert
yields inside iterations over _bucket in other subsystems to do
not occupy TX thread too long when _bucket is huge.

The patch makes this generation be global and available
everywhere in the module.