gc: make bucket GC tolerant to vinyl
Bucket GC is a vshard subsystem originaly designed to clear
and delete sent and garbage buckets out of a storage. But it was
not able to work with vinyl because it was not aware of yielding
selects and multi-engine transactions.

Now GC is reworked almost completely to get on with vinyl and has
some non-related to vinyl improvements:
    * it deletes buckets and bucket data in batches, not in a
      huge transactions that even could not fit in memory;
    * bucket generation checking now is more accurate: each
      bucket is checked individually instead of global GC restart
      on any _bucket change.

Closes #132
Part of #85