rebalancer: make rebalancer be tolerant to vinyl
Before the patch the main problem withstanding Vinyl space
rebalancing was the fact that each bucket was being applied on
a destination replicaset as a huge monolite transaction, failing
due to lack of multiengine transactions. Memtx _bucket + any
vinyl space are not compatible.

The patch makes bucket applying more grained. A bucket applying
now consists of at least 3 transactions: make the bucket be
RECEIVING, apply the data, make the bucket be ACTIVE. Data
appling is splitted into multiple transactions in a common case:
each applies data of one space and at most 1000 tuples per time.

In turn the recovery procedure is slightly changed because before
the patch RECEIVING bucket status actually was unreachable. Now
it is not so, and such buckets should be accounted by the
recovery more carefully considering the most complex case: when a
bucket is SENDING on source and RECEIVING on destination. This
situation can be resolved in two ways depending on rebalancer
status. If the bucket is not actually sending now (for example,
an instance has restarted during sending/receiving), then the
sending has failed and the bucket is activated on the source
again. If the bucket sending is in progress still (for example,
the bucket is big and its receiving consists of multiple
transactions), then the recovery does not try to repair it while.

Part of #73