rebalancer: send buckets in parts
It is possible that a cluster is huge, consists of significant
number or replicasets, has terrible bucket count near 500k ~ 1kk,
and each bucket has size of tens of megabytes or more.

Such bucket should not be sent in a single cosmic message,
otherwise its applying will linger on. A source replicaset has
to split the bucket into multiple messages of size about 1k
tuples.

Besides, this patch is one step forward to non-blocking the
bucket for the whole transfer interval.

Part of #73