recovery: start recovery from destination always
Consider the case: a last bucket message is sent to a
destination but hung in the network. Then the connection is
re-established and the recovery started working. It recovers
the bucket into active state on the source. Then the hung
message arrives to the destination, applies data and the bucket
is activated on the second storage.

To prevent it do a couple of small but important things:
* never send a bucket in a single message - in such a case there
  is always a possibility, that this message hangs and the bucket
  is activated on the source. And there is no way to detect a
  hang reason. So actually a bucket is sent now in two phases:
  prepare (consists from >= 1 messages) and commit.

* on recovery at first turn a junk bucket into garbage state, and
  only then turn the initial bucket into active state. In such a
  case even if a hung message arrived after the bucket is already
  recovered, it will be just rejected.

Closes #149