storage: return to fullmesh topology
Fullmesh was disabled by #66 because of its too raw
implementation in Tarantool. Now it is ripe enough to be
granted a second chance.

Fullmesh provides a benefit of a constant complete sync. If a
tuple is written on a master right before its demotion, a new
master still will receive the tuple from the old one.

Closes #83