recovery: simplify the procedure of recovery
Before the patch a special map of buckets to recovery was
maintained during a master lifecycle. But 1) the same buckets
can be got from _bucket.index.status iterator, 2) it does not
work when a master switch emerges during transfer so a bucket
arrives to a new master via replication after it is configured.
Such buckets are not put into the map of buckets to recovery.

This patch gets rid of this map and uses directly
_bucket.index.status.