router: allow to pass any netbox options to call
Before the patch only timeout was available for
setting by a user. But tempora mutantur, and some
users want to be able to use box.session.push() and
netbox is_async. This patch allows any netbox options.
This is enough for pushes.

Closes #82

But not enough for netbox is_async because
vshard.router.call calls a user function not directly
but via vshard.storage.call.

Needed for #171