router: support netbox is_async flag
Previous commit enables full support of netbox options
in router.call and replicaset.call. But it was not
enough for async requests because vshard.router.call
calls a user function not directly but via
vshard.storage.call wrapper which checks bucket
availability.

So vshard.router.call relied on two facts:

* first returned value from netbox.call is true/false
  as described by storage.call;

* actually returned value is on a second position.

Example:

    status, value = vshard.storage.call(func)

Here status - true/false. Value is a returned value
of func(), if status is true.

So vshard.router.call needed a special support of
async requests to be aware that async returns not
true/false and a result, but future object and an
error. Also a special unpacker was needed to extract
a func returned value from a future.result.

This patch makes vshard.router.call aware about async
requests, and wraps its future object with a special
unpacker so as future objects behave the same as
original ones.

Closes #171