router: do not close an old connection on failover
Sometimes it happens that a router processed a huge
response. Processing can take more than failover ping
timeout, and in such a case failover's pings fail. But
the connection is alive. So instead of closing it,
this commit detaches the connection from its replica
object. Existing requests can finish their job with this
connection.

Closes #169