router: create a template function for RO requests
At this moment there is only one RO request - call on a nearest
replica or master otherwise. But #157 and #168 are going to add
more: prefer slaves for a call; do load-balancing over all; do
load-balancing, but prefer slaves.

It would be easier if all of those features could be implemented
via one customized template.