router: introduce router.callre
vshard.router.callre is a method similar to callro, but it
prefers non-master nodes - slaves. If there are alive slaves,
this method will call the function on one of them, prefering a
one with the highest priority. Otherwise the master will be used.

Closes #157

@TarantoolBot document
Title: [vshard] document router.callre, new vshard.call, and replicaset:callre
VShard router supports slave-preference calls. It is a call
executed on the nearest active slave node. If there are no ready
at hand alive slave, then the call is executed on master.

To execute such a call use vshard.router.callre, or
replicaset:callre. Arguments are the same as for
vshard.router.callro and replicaset:callro.

Also, vshard.router.call is updated. Now it supports two sets of
arguments. The first is old:

    vshard.router.call(bucket_id, mode, func, args, opts)

The second is new:

    vshard.router.call(bucket_id, shard_opts, func, args,
                       net_opts)

Shard opts is a map with the following keys: mode = boolean, and
prefer_replica = boolean.

Why could a one need such a strange thing as replica preference?
For example, if a call is heavy in terms of disk access or
calculations. Then it would be better to do not bother a master
node with that.