router: fix a bug with lost second return value
The template function, being used to generate router call
versions with different flags, accidentally had declared one
local variable two times - inside a 'while' loop and out of it.
Inside the loop it was assigned a second returned value, but it
was returning the outer version, being always nil. The commit
fixes that obvious bug.