storage: on reload do not forget about FFI ref constructor
VShard storage has a global variable - bucket_ref_new(), it is an
FFI function to create struct bucket_ref. But it can not be
created again using ffi.metatype(), even on reload, and it was
just nullified. This patch resets it using ffi.typeof() on
reload.

Thanks Vladislav Grubov for the proposed solution.

Closes #178