storage: fix vshard.storage.sync() on stopped downstream
vshard.storage.sync() is a function to sync this instance with
its replicas, so instance's LSN is learned by the replicas.

It was quite simple - check downstream vclock, and return ok,
when all downstream vlocks are >= this. But the thing is that
'box.info.replication[...].downstream.vclock' field disappears,
when connection is down.

Now sync considers stopped replicas as an error.

Closes #180