rebalancer: log maximal disbalance
It happens that users see in a log: "The cluster is balanced ok",
but nothing happens even when obviously there is no balance. For
example, two replicasets have weights 100/150 and bucket counts
10000/20000. Perfect balance is 12000/18000 with such weights,
but nothing changes.

In that concrete case appeared, that the customer had 30%
disbalance threshold in the configuration. To avoid such
misunderstanding this patch adds more detailed logging in case
there is disbalance but nothing changes.

Closes #191