rebalancer: introduce dispenser
The rebalancer will become parallel soon - it will be able to
send multiple buckets from one node simultaneously. For that it
will create several worker fibers. But they need somehow
synchronize their work so as not to send too many buckets, or too
few, and to spread the load evenly among receivers.

Dispenser is a container of routes received from the rebalancer.
Its task is to hand out the routes to worker fibers in a
round-robin manner. It solves all the problems listed above.

Part of #161