gc: fix error about double deletion of bucket
Garbage collector was trying to drop 'sent' buckets multiple
times. This was happening because the collector keeps two arrays
of 'sent' buckets: one is scheduled for deletion after a timeout,
and another is kept to be scheduled next.

'Sent' buckets are not deleted immediately so as to allow routers
learn a new location of the bucket. Two arrays allow to make that
process more step-wise, and never delete a sent bucket before the
timeout.

But that algorithm led to a problem, that some buckets were
accounted in both arrays. Indeed, if a bucket is 'sent' and is
already scheduled for deletion, nonetheless it is still in
_bucket space, and is added to the second array too. That made
bucket guard angry.

The patch gets rid of bucket guard here and manually allows to
skip already deleted buckets.