storage: move control over max receiving to receivers
Rebalancer algorithm is quite a complex thing, and it is better
not to complicate it even more with the max receiving bucket
count restriction. Now all the control is done on receiver's
side.

At this moment complexity of rebalancer algo is not big, but it
relies on a fact, that each replicaset can send only one bucket
at once. It won't be true after #161, and this is where the
algorithm would become unbearably complex.

Closes #167