router: fix reload problem with global function refs
During reload even without code changes all global functions
change their addresses. That is, a <func_name> before reload and
<func_name> after reload at '==' will return false. But there
was a place which compared functions by addresses - triggers.

Triggers are implemented as a list of functions from which it is
possible to remove old and add new ones. Removal of an old throws
an exception when nothing is found. So if code adds a global
function to a trigger list, does reload, and tries to remove it
by even the same name, it will get the exception.

To protect from reload all such places should save a function
reference somewhere to be able to remove a trigger by that
reference instead of a global name.

Closes #193