Configuration checker should not treat 'false' as 'nil'
Config validator used 'not value' to check if 'value' is not set.
But 'value' can be 'false', and it should not be treated as an
error.

This didn't break anything earlier, because all boolean values
either were optional and had 'false' default value, or had a
custom validator-function.

This would not break anything even now, but a bug is a bug.