router: bootstrap() if_not_bootstrapped option
Closes #204

@TarantoolBot document
Title: bootstrap() if_not_bootstrapped option
`vshard.router.bootstrap()` used to accept only one optional
argument `timeout`: `vshard.router.bootstrap({timeout = ...})`.
And it was returning an error, when discovered that the cluster is
already bootstrapped.

It is not useful for most of the clients, because usually they
call `vshard.router.bootstrap()` every time when a router is
started. This is of course a bad practice!!!, but nonetheless
there were many requests for an option not to raise an error, when
the cluster is already bootstrapped.

That option is added and is called `if_not_bootstrapped`. By
default it is `false`. `False` means "raise an error", this is old
behaviour. `True` means consider an already bootstrapped cluster a
success. Then `true` is returned.

Note, that bootstrapped a cluster or not is detected by looking
for at least one bucket in the whole cluster. If the cluster was
bootstrapped only partially (for example, due to an error during
the first bootstrap), then it will be considered a bootstrapped
cluster on a next bootstrap call with `if_not_bootstrapped`. So
this is still a bad practice. Avoid calling `bootstrap()` multiple
times.

Usage example:
`vshard.router.bootstrap({timeout = ..., if_not_bootstrapped = true})`