router: not replica:is_connected() != no connection
replica:is_connected() is a shortcut for
replica.conn:is_connected(). Conn is a netbox object. It says
is_connected() true, when the TCP connection is established.

It may happen, that TCP is ok, but netbox on_connect() triggers
are not fired yet (a connection needs to fetch a schema, and do
auth). It was not taken into account in the failover fiber, which
couldn't understand, why a connection still has down timestamp
not cleared, but netbox says that it is connected.

Now the problem is covered. The failover fiber considers a
connection ok, if it has down timestamp not cleared, but TCP is
ok. Because it is known for sure the connection will clear the
down timestamp soon, after schema fetch and auth are finished.

Closes #205