rebalancer: don't activate bucket on sender first
A very nasty bug. Here is what happened. There are 4 replicasets:
rs1, rs2, rs3, rs4.

Rs1 gets weight 0. It sends buckets to rs2, rs3, rs4 in parallel.
Sending of each bucket worked in two steps: send data, and
finalize bucket making it ACTIVE on receiving side. If an error
happened before sending finalization, the bucket became ACTIVE on
the sender immediately. There was an assumption, that if the final
packet was not sent, the bucket, indeed, couldn't be ACTIVE on the
receiving side.

As a result, rs1 partially sent a bucket to rs2, declared it
ACTIVE again, and successfully sent it to rs3 or rs4. Rs2 recovery
woke up and saw, that the bucket is RECEIVING here and GARBAGE/not
existing in rs1. As a result, it was recovered to ACTIVE on rs2
too. Now the bucket is ACTIVE on rs2 and rs3 or rs4.

On the summary, this broke recovery rules. According to them,
there is a strict order how a bucket should be recovered - firstly
RECEIVING should be garbaged, and only then the SENDING should
become active again.

The patch removes the immediate activation optimization and lets
the recovery subsystem do its job.

Closes #213