test: test reload to the latest version
Reload_evolution test is a test on how hot code reload works. And
it didn't test the latest version of vshard. It was checking
reload from version

    b9b63a8b2c473b644519065c77bf77f4c26beab9
    Date:   Thu Aug 2 13:27:10 2018 +0300
    Refactor config templates

to

    f3246217aab7f33d98375cbf42c0fb3a919c6ba3
    Date:   Fri Jun 29 20:34:26 2018 +0300
    Introduce storage reload evolution

It was close to useless basically. After the test was fixed, it
appeared the hot code reload was broken a couple of times already.
For example, bucket_refs introduction was not reflected in the
reload handlers. And such a hard code change anyway could not be
reloaded successfully without restart.

This commit finds an old enough point in the past, from where
reload should work fine, and makes the reload_evolution test
check reload from that version to the latest.