test: print errors in a portable way
In 2.4.1 error objects have new 'base_type' field. That makes
number of unstable fields 2: 'trace' and 'base_type'. The latter
is unstable, because does not exist before 2.4.1.

The patch introduces a function which trims all not interesting
fields from error objects in the tests. That makes the test work
on all supported tarantools.