test: disable router discovery for some tests
There are tests, which are not related to discovery at all, and
yet they suffer from unrelated changes, when something is modified
in the discovery algorithm. For example, it highly affects the
test about request retries and automatic timeouts.

Now the discovery is disabled for such tests, since it becomes
easily possible to do with the new router's configuration option.