router: fix missing constant declaration
Router for tarantool 1.9 used a not declared variable
DISCOVERY_INTERVAL. It was replaced with a new set of DISCOVERY_*
variables for >= 1.10 in 3c16527fc30d5d0910e5a4675775333ffb8c66d7
("router: make discovery smoother in a big cluster").

The patch makes it use DISCOVERY_IDLE_INTERVAL. Because in 1.9
discovery is not stepwise.

Closes #250