util: introduce binary heap data structure
Lua does not have a built-in standard library for binary heaps
(also called priority queues). There is an implementation in
Tarantool core in libsalad, but it is in C.

Heap is a perfect storage for the soon coming feature map-reduce.
In the map-reduce algorithm it will be necessary to be able to
lock an entire storage against any bucket moves for time <=
specified timeout. Number of map-reduce requests can be big, and
they can have different timeouts.

So there is a pile of timeouts from different requests. It is
necessary to be able to quickly add new ones, be able to delete
random ones from any place, and remove expired ones.

One way would be a sorted array of the timeouts. Unfortunately,
it is super slow. O(N * log(N)) to add a new element (find place
for log(N) and move all next elements for N), O(N) to delete a
random one (move all next elements one cell left/right).

Another way would be a sorted tree. But trees like RB or a dumb
binary tree require extra steps to keep them balanced and to have
access to the smallest element ASAP.

The best way is the binary heap. It is perfectly balanced by
design meaning that all operations there have complexity at most
O(log(N)). It is possible to find the closest deadline for
constant time as it is the heap's top.

This patch implement it. The heap is intrusive. It means it stores
index of each element right inside of the element as a field
'index'. Having an index along with each element allows to delete
it from the heap for O(log(N)) without necessity to find its place
first.

Heap can be sorted as max or min. For ordering it uses plain
comparison operators which are expected to be overloaded for the
stored type.

Part of #147